kubelet: Clear the podStatusChannel before invoking syncBatch https://github.com/kubernetes/kubernetes/pull/88440/commits/8bc5cb01a9c1e7698ddd641963da08d09410dcfc
The status manager syncBatch() method processes the current state
of the cache, which should include all entries in the channel. Flush
the channel before we call a batch to avoid unnecessary work and
to unblock pod workers when the node is congested.

Discovered while investigating long shutdown intervals on the node
where the status channel stayed full for tens of seconds.

Add a for loop around the select statement to avoid unnecessary
invocations of the wait.Forever closure each time.

committed on Mar 4, 2020