Switch the namespace controller to use the metadata client https://github.com/kubernetes/kubernetes/pull/78744/commits/50fd47258d97299a34b41c98a0ce23bcceee2326 committed on Jul 10, 2019
The metadata client uses protobuf and returns only a subset of object
data (the metadata) which allows operations that act only on objects
generically to work much faster. Use the metadata client in the
namespace controller to reduce the amount of work the namespace controller
has to do in large namespaces.

# Conflicts:
#	cmd/kube-controller-manager/app/core.go
#	pkg/controller/namespace/deletion/namespaced_resources_deleter.go
#	pkg/controller/namespace/deletion/namespaced_resources_deleter_test.go
#	pkg/controller/namespace/namespace_controller.go
#	test/e2e_node/services/namespace_controller.go