Always negotiate a decoder using ClientNegotiator https://github.com/kubernetes/kubernetes/pull/84692/commits/3b780c64b89606f4e6b21f48fb9c305d5998b9e5
This commit performs two refactors and fixes a bug.

Refactor 1 changes the signature of Request to take a RESTClient, which
removes the extra copy of everything on RESTClient from Request. A pair
of optional constructors are added for testing. The major functional
change is that Request no longer has the shim HTTPClient interface and
so some test cases change slightly because we are now going through
http.Client code paths instead of direct to our test stubs.

Refactor 2 changes the signature of RESTClient to take a
ClientContentConfig instead of ContentConfig - the primary difference
being that ClientContentConfig uses ClientNegotiator instead of
NegotiatedSerializer and the old Serializers type. We also collapse
some redundancies (like the rate limiter can be created outside the
constructor).

The bug fix is to negotiate the streaming content type on a Watch()
like we do for requests. We stop caching the decoder and simply
resolve it on the request. We also clean up the dynamic client
and remove the extra WatchSpecificVersions() method by providing
a properly wrapped dynamic client.

# Conflicts:
#	staging/src/k8s.io/client-go/metadata/fake/simple.go
#	staging/src/k8s.io/client-go/rest/config_test.go
#	staging/src/k8s.io/client-go/rest/request.go
#	staging/src/k8s.io/client-go/rest/request_test.go