Master 822 missing cherry pick apiserver (#949)
* flowcontrol context aware and fix request may hang issue
https://github.com/kubernetes/kubernetes/pull/79375/commits/ce8805f95fcf6540397eaa60b8d84db752f05eea
answer1991 committed on Jun 25, 2019
pick up for https://github.com/kubernetes/kubernetes/pull/84692 dependency

* Creating function for preflight check.
https://github.com/kubernetes/kubernetes/pull/82652/commits/ad5fafd6ade2838098890a4e7727c8e347686867
For clean 84692

Migrated code that checks for common programmer errors to a separated
function and added test coverage for it. Wrong comment stating that a
typed error is returned was also removed.

* Remove deprecated-dynamic client
https://github.com/kubernetes/kubernetes/pull/84692/commits/c416ee584c178bb89c6cd11c93b504f2098fac0f
smarterclayton committed on Nov 10, 2019

* Detect watch protocol errors via an e2e test for apimachinery
https://github.com/kubernetes/kubernetes/pull/84692/commits/15f5e64404e99d1dcfa3bfc46ca4441443cc2904
committed on Nov 10, 2019
part of https://github.com/kubernetes/kubernetes/pull/84692

This e2e test reproduces #62175 and will be expanded to check for other
negotiation related errors against real servers.

* test: Watch should fail immediately on negotiate errors
https://github.com/kubernetes/kubernetes/pull/84692/commits/9aad6aa54d824ba93a6670cd5a0cab6ad337e9f0
part of https://github.com/kubernetes/kubernetes/pull/84692

Instead of returning an error on the watch stream, if we can't properly
negotiate a watch serialization format we should error and return that
error to the client.

* dynamic: The dynamic client no longer needs a special cased watch
https://github.com/kubernetes/kubernetes/pull/84692/commits/3f94f80b0a79293e54d7080aaf7a64d7df8b1d4a
part of https://github.com/kubernetes/kubernetes/pull/84692

By correctly handling content type negotiation, we can avoid the
need for a special version of watch and use the same code path as
typed clients.

# Conflicts:
#	staging/src/k8s.io/client-go/dynamic/simple.go

* test: Set RateLimiter via client config vs direct casting
https://github.com/kubernetes/kubernetes/pull/84692/commits/0ba0ef057a494e9c77a4fbb36b76096bab9fd277
part of https://github.com/kubernetes/kubernetes/pull/84692

This test has no need to perform a dereference to an implementation,
it can instead set the rate limited during initialization.

* test: kubectl unit tests should be using codecs without conversion
https://github.com/kubernetes/kubernetes/pull/84692/commits/8a9b8c87c40ee65751828d9dd02f4f642588f0ce

Tests are also refactored to use the simpler RESTClient code path.

# Conflicts:
#	staging/src/k8s.io/kubectl/pkg/cmd/attach/attach_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_clusterrolebinding_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_configmap_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_deployment_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_namespace_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_pdb_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_priorityclass_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_rolebinding_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_secret_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/create/create_serviceaccount_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/drain/drain_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/exec/exec_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/expose/expose_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/portforward/portforward_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/rollout/BUILD
#	staging/src/k8s.io/kubectl/pkg/cmd/rollout/rollout_pause_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/run/run_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/taint/taint_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/top/top_node_test.go
#	staging/src/k8s.io/kubectl/pkg/cmd/top/top_pod_test.go

* test: Exit early during resource helper test
https://github.com/kubernetes/kubernetes/pull/84692/commits/b4531067770645bfeb64fa88a8ced449313597af

The test will panic if it fails, and should instead check and exit
when invalid conditions are hit.

* Copyright fix for cherry pick 1.18 + bazel update

* Always negotiate a decoder using ClientNegotiator
https://github.com/kubernetes/kubernetes/pull/84692/commits/3b780c64b89606f4e6b21f48fb9c305d5998b9e5

This commit performs two refactors and fixes a bug.

Refactor 1 changes the signature of Request to take a RESTClient, which
removes the extra copy of everything on RESTClient from Request. A pair
of optional constructors are added for testing. The major functional
change is that Request no longer has the shim HTTPClient interface and
so some test cases change slightly because we are now going through
http.Client code paths instead of direct to our test stubs.

Refactor 2 changes the signature of RESTClient to take a
ClientContentConfig instead of ContentConfig - the primary difference
being that ClientContentConfig uses ClientNegotiator instead of
NegotiatedSerializer and the old Serializers type. We also collapse
some redundancies (like the rate limiter can be created outside the
constructor).

The bug fix is to negotiate the streaming content type on a Watch()
like we do for requests. We stop caching the decoder and simply
resolve it on the request. We also clean up the dynamic client
and remove the extra WatchSpecificVersions() method by providing
a properly wrapped dynamic client.

# Conflicts:
#	staging/src/k8s.io/client-go/metadata/fake/simple.go
#	staging/src/k8s.io/client-go/rest/config_test.go
#	staging/src/k8s.io/client-go/rest/request.go
#	staging/src/k8s.io/client-go/rest/request_test.go

* client config ut test fix for cherry-pick from 1.18 + Copyright fix