1.18 improvement kubelet (#931)
* Fix sync pod log format and a func typo.
https://github.com/kubernetes/kubernetes/pull/82113/commits/8dc401d14161e757540ab46e40ed443d0300cdb5
committed on Aug 28, 2019

# Conflicts:
#	pkg/util/pod/pod.go

* Add UID precondition to kubelet pod status patch updates
https://github.com/kubernetes/kubernetes/pull/86320/commits/a65d8aeb7622a56b373f9d5c3762c07a4b776d21
committed on Dec 16, 2019

# Conflicts:
#	pkg/kubelet/status/status_manager.go
#	pkg/util/pod/pod.go
#	pkg/util/pod/pod_test.go

* kubelet: Avoid sending no-op patches
https://github.com/kubernetes/kubernetes/pull/88591/commits/b2528654797e209f74dd9c9267e205098b4092c0
committed on Feb 26, 2020

In an e2e run, out of 1857 pod status updates executed by the
Kubelet 453 (25%) were no-ops - they only contained the UID of
the pod and no status changes. If the patch is a no-op we can
avoid invoking the server and continue.

# Conflicts:
#	pkg/kubelet/status/status_manager.go
#	pkg/util/pod/pod.go
#	pkg/util/pod/pod_test.go

* Fix API reference changes from cherry pick

* Test that an always-fail container can't report the pod Succeeded
https://github.com/kubernetes/kubernetes/pull/88440/commits/6d98b0a0f4fbb9c1032f5c47fa20f59489301138
The kubelet can race when a pod is deleted and report that a container succeeded
when it instead failed, and thus the pod is reported as succeeded. Create an e2e
test that demonstrates this failure.
committed on Mar 4, 2020

# Conflicts:
#	test/e2e/node/pods.go

Pick up from https://github.com/kubernetes/kubernetes/pull/88440/commits
Ensure Kubelet always reports terminating pod container status #88440

* Fix cherry-pick error

* kubelet: Preserve existing container status when pod terminated
https://github.com/kubernetes/kubernetes/pull/88440/commits/ad3d8949f08766106c203e273b30c810994ebd99

The kubelet must not allow a container that was reported failed in a
restartPolicy=Never pod to be reported to the apiserver as success.
If a client deletes a restartPolicy=Never pod, the dispatchWork and
status manager race to update the container status. When dispatchWork
(specifically podIsTerminated) returns true, it means all containers
are stopped, which means status in the container is accurate. However,
the TerminatePod method then clears this status. This results in a
pod that has been reported with status.phase=Failed getting reset to
status.phase.Succeeded, which is a violation of the guarantees around
terminal phase.

Ensure the Kubelet never reports that a container succeeded when it
hasn't run or been executed by guarding the terminate pod loop from
ever reporting 0 in the absence of container status.
committed on Mar 4, 2020

* kubelet: Don't delete pod until all container status is available
https://github.com/kubernetes/kubernetes/pull/88440/commits/2364c10e2e4d9fc04e565f2e07f400e19b1bcc41

After a pod reaches a terminal state and all containers are complete
we can delete the pod from the API server. The dispatchWork method
needs to wait for all container status to be available before invoking
delete. Even after the worker stops, status updates will continue to
be delivered and the sync handler will continue to sync the pods, so
dispatchWork gets multiple opportunities to see status.

The previous code assumed that a pod in Failed or Succeeded had no
running containers, but eviction or deletion of running pods could
still have running containers whose status needed to be reported.

This modifies earlier test to guarantee that the "fallback" exit
code 137 is never reported to match the expectation that all pods
exit with valid status for all containers (unless some exceptional
failure like eviction were to occur while the test is running).

committed on Mar 4, 2020

* kubelet: Never restart containers in deleting pods
https://github.com/kubernetes/kubernetes/pull/88440/commits/8722c834e5e52e241c0227e8673837c9081c1423

When constructing the API status of a pod, if the pod is marked for
deletion no containers should be started. Previously, if a container
inside of a terminating pod failed to start due to a container
runtime error (that populates reasonCache) the reasonCache would
remain populated (it is only updated by syncPod for non-terminating
pods) and the delete action on the pod would be delayed until the
reasonCache entry expired due to other pods.

This dramatically reduces the amount of time the Kubelet waits to
delete pods that are terminating and encountered a container runtime
error.

committed on Mar 4, 2020

* kubelet: Clear the podStatusChannel before invoking syncBatch
https://github.com/kubernetes/kubernetes/pull/88440/commits/8bc5cb01a9c1e7698ddd641963da08d09410dcfc

The status manager syncBatch() method processes the current state
of the cache, which should include all entries in the channel. Flush
the channel before we call a batch to avoid unnecessary work and
to unblock pod workers when the node is congested.

Discovered while investigating long shutdown intervals on the node
where the status channel stayed full for tens of seconds.

Add a for loop around the select statement to avoid unnecessary
invocations of the wait.Forever closure each time.

committed on Mar 4, 2020