1.18 improvement kcm metadataclient (#935)
* Add a metadata client to client-go that can read PartialObjectMetadata
https://github.com/kubernetes/kubernetes/pull/77819/commits/21f5e643d9dbe6b65d21713dc16ab8888de5423e
committed on Jul 2, 2019

This client exposes operations on generic metadata (get, list, watch, delete)
and allows patch operations. The client always uses protobuf and requests
the server transform the response into the appropriate object. Using this
client simplifies the work of generic controllers by allowing them to treat
all objects the same, and also improves performance both in the amount of
data sent as well as allowing protobuf on CRD resources.

* Fake ObjectReaction should handle PartialObjectMetadata special
https://github.com/kubernetes/kubernetes/pull/78744/commits/baf091e9dbad00db39e246815f9d7a21d148044f
committed on Jul 10, 2019

When a client requests a PartialObjectMetadata returned from the
ObjectReaction type, if the object has a GVK set use that instead
of what the schema returns, since the majority of clients getting
partial object metadata will be doing so using the metadata client
or server side conversion.

* Add fake client, informer factory, and lister to metadata client
https://github.com/kubernetes/kubernetes/pull/78744/commits/bc89c37f32aa6cfd0f9ca975d9221d0a89320623
committed on Jul 10, 2019

These will be used by the garbage collector controller and others that
use higher level primitives.

* Switch the namespace controller to use the metadata client
https://github.com/kubernetes/kubernetes/pull/78744/commits/50fd47258d97299a34b41c98a0ce23bcceee2326
committed on Jul 10, 2019

The metadata client uses protobuf and returns only a subset of object
data (the metadata) which allows operations that act only on objects
generically to work much faster. Use the metadata client in the
namespace controller to reduce the amount of work the namespace controller
has to do in large namespaces.

# Conflicts:
#	cmd/kube-controller-manager/app/core.go
#	pkg/controller/namespace/deletion/namespaced_resources_deleter.go
#	pkg/controller/namespace/deletion/namespaced_resources_deleter_test.go
#	pkg/controller/namespace/namespace_controller.go
#	test/e2e_node/services/namespace_controller.go

* Switch the garbage collector to use metadata client and protobuf
https://github.com/kubernetes/kubernetes/pull/78742/commits/637cc83341f7e8cb80eb974f614206c2112b2ad7
committed on Jul 11, 2019

# Conflicts:
#	pkg/controller/.import-restrictions
#	pkg/controller/garbagecollector/garbagecollector.go
#	pkg/controller/garbagecollector/operations.go

* Add multi-tenancy changes for metadataclient. Fix unit tests, integration test, copyrights, make update